Split model regions out into separate files
