+++ src/ModelRegion.cpp
@@ -30,50 +30,21 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+#include "config.h"
 #include "ModelRegion.hpp"
 
-#include <stdlib.h>
-#include <iostream>
-#include <time.h>
-#include <cmath>
-#include <string>
-#include <string.h>
-#include <mpi.h>
-#ifdef _OPENMP
-#include <omp.h>
-#endif
-
 #include "geopm.h"
-#include "geopm_time.h"
+#include "geopm_imbalancer.h"
 #include "Exception.hpp"
 #include "Helper.hpp"
-#include "geopm_imbalancer.h"
-#include "config.h"
 
-#ifdef GEOPM_ENABLE_MKL
-#include <mkl.h>
-#else
-// Terrible DGEMM implementation should only be used if there is no
-// BLAS support.  Build assumes that the Intel(R) Math Kernel Library
-// is the only provider of BLAS.
-static inline
-void dgemm(const char *transa, const char *transb, const int *M,
-           const int *N, const int *K, const double *alpha,
-           const double *A, const int *LDA, const double *B,
-           const int *LDB, const double *beta, double *C, const int *LDC)
-{
-#pragma omp parallel for
-    for (int i = 0; i < *M; ++i) {
-        for (int j = 0; j < *N; ++j) {
-            C[i * *LDC + j] = 0;
-            for (int k = 0; k < *K; ++k) {
-                C[i * *LDC + j] += A[i * *LDA + j] * B[j * *LDB + k];
-            }
-        }
-    }
-}
-#endif
-
+#include "NestedModelRegion.hpp"
+#include "SleepModelRegion.hpp"
+#include "All2allModelRegion.hpp"
+#include "DGEMMModelRegion.hpp"
+#include "StreamModelRegion.hpp"
+#include "SpinModelRegion.hpp"
+#include "IgnoreModelRegion.hpp"
 
 namespace geopm
 {
@@ -89,8 +60,9 @@ namespace geopm
         return result;
     }
 
-
-    std::unique_ptr<ModelRegion> ModelRegion::model_region(const std::string &name, double big_o, int verbosity)
+    std::unique_ptr<ModelRegion> ModelRegion::model_region(const std::string &name,
+                                                           double big_o,
+                                                           int verbosity)
     {
         bool do_imbalance = (name.find("-imbalance") != std::string::npos);
         bool do_progress = (name.find("-progress") != std::string::npos);
@@ -210,555 +182,6 @@ namespace geopm
     {
         if (m_do_imbalance) {
             (void)geopm_imbalancer_exit();
-        }
-    }
-
-    SleepModelRegion::SleepModelRegion(double big_o_in, int verbosity, bool do_imbalance, bool do_progress, bool do_unmarked)
-        : ModelRegion(verbosity)
-    {
-        m_name = "sleep";
-        m_do_imbalance = do_imbalance;
-        m_do_progress = do_progress;
-        m_do_unmarked = do_unmarked;
-        big_o(big_o_in);
-        int err = ModelRegion::region();
-        if (err) {
-            throw Exception("SleepModelRegion::SleepModelRegion()",
-                            err, __FILE__, __LINE__);
-        }
-    }
-
-    SleepModelRegion::~SleepModelRegion()
-    {
-
-    }
-
-    void SleepModelRegion::big_o(double big_o_in)
-    {
-        num_progress_updates(big_o_in);
-        double seconds = big_o_in / m_num_progress_updates;
-        m_delay = {(time_t)(seconds),
-                   (long)((seconds - (time_t)(seconds)) * 1E9)};
-
-        m_big_o = big_o_in;
-    }
-
-    void SleepModelRegion::run(void)
-    {
-        if (m_big_o != 0.0) {
-            if (m_verbosity) {
-                std::cout << "Executing " << m_big_o << " second sleep."  << std::endl << std::flush;
-            }
-            ModelRegion::region_enter();
-            for (uint64_t i = 0 ; i < m_num_progress_updates; ++i) {
-                ModelRegion::loop_enter(i);
-                int err;
-                err = clock_nanosleep(CLOCK_REALTIME, 0, &m_delay, NULL);
-                if (err) {
-                    throw Exception("SleepModelRegion::run()",
-                                    GEOPM_ERROR_RUNTIME, __FILE__, __LINE__);
-                }
-                ModelRegion::loop_exit();
-            }
-            ModelRegion::region_exit();
-        }
-    }
-
-    SpinModelRegion::SpinModelRegion(double big_o_in, int verbosity, bool do_imbalance, bool do_progress, bool do_unmarked)
-        : ModelRegion(verbosity)
-    {
-        m_name = "spin";
-        m_do_imbalance = do_imbalance;
-        m_do_progress = do_progress;
-        m_do_unmarked = do_unmarked;
-        big_o(big_o_in);
-        int err = ModelRegion::region();
-        if (err) {
-            throw Exception("SpinModelRegion::SpinModelRegion()",
-                            err, __FILE__, __LINE__);
-        }
-    }
-
-    SpinModelRegion::~SpinModelRegion()
-    {
-
-    }
-
-    void SpinModelRegion::big_o(double big_o_in)
-    {
-        num_progress_updates(big_o_in);
-        m_delay = big_o_in / m_num_progress_updates;
-        m_big_o = big_o_in;
-    }
-
-    void SpinModelRegion::run(void)
-    {
-        if (m_big_o != 0.0) {
-            if (m_verbosity) {
-                std::cout << "Executing " << m_big_o << " second spin."  << std::endl << std::flush;
-            }
-            ModelRegion::region_enter();
-            for (uint64_t i = 0 ; i < m_num_progress_updates; ++i) {
-                ModelRegion::loop_enter(i);
-
-                double timeout = 0.0;
-                struct geopm_time_s start = {{0,0}};
-                struct geopm_time_s curr = {{0,0}};
-                (void)geopm_time(&start);
-                while (timeout < m_delay) {
-                    (void)geopm_time(&curr);
-                    timeout = geopm_time_diff(&start, &curr);
-                }
-
-                ModelRegion::loop_exit();
-            }
-            ModelRegion::region_exit();
-        }
-    }
-
-    DGEMMModelRegion::DGEMMModelRegion(double big_o_in, int verbosity, bool do_imbalance, bool do_progress, bool do_unmarked)
-        : ModelRegion(verbosity)
-        , m_matrix_a(NULL)
-        , m_matrix_b(NULL)
-        , m_matrix_c(NULL)
-        , m_matrix_size(0)
-        , m_pad_size(64)
-    {
-        m_name = "dgemm";
-        m_do_imbalance = do_imbalance;
-        m_do_progress = do_progress;
-        m_do_unmarked = do_unmarked;
-        big_o(big_o_in);
-        int err = ModelRegion::region(GEOPM_REGION_HINT_COMPUTE);
-        if (err) {
-            throw Exception("DGEMMModelRegion::DGEMMModelRegion()",
-                            err, __FILE__, __LINE__);
-        }
-    }
-
-    DGEMMModelRegion::~DGEMMModelRegion()
-    {
-        big_o(0);
-    }
-
-    void DGEMMModelRegion::big_o(double big_o_in)
-    {
-        if (m_big_o && m_big_o != big_o_in) {
-            free(m_matrix_c);
-            free(m_matrix_b);
-            free(m_matrix_a);
-        }
-
-        num_progress_updates(big_o_in);
-
-        m_matrix_size = (int)pow(4e9 * big_o_in / m_num_progress_updates, 1.0/3.0);
-        if (big_o_in && m_big_o != big_o_in) {
-            size_t mem_size = sizeof(double) * (m_matrix_size * (m_matrix_size + m_pad_size));
-            int err = posix_memalign((void **)&m_matrix_a, m_pad_size, mem_size);
-            if (!err) {
-                err = posix_memalign((void **)&m_matrix_b, m_pad_size, mem_size);
-            }
-            if (!err) {
-                err = posix_memalign((void **)&m_matrix_c, m_pad_size, mem_size);
-            }
-            if (err) {
-                throw Exception("DGEMMModelRegion::big_o(): posix_memalign() failed",
-                                err, __FILE__, __LINE__);
-            }
-#pragma omp parallel for
-            for (size_t i = 0; i < mem_size / sizeof(double); ++i) {
-                m_matrix_a[i] = 2.0 * i;
-                m_matrix_b[i] = 3.0 * i;
-            }
-        }
-        m_big_o = big_o_in;
-    }
-
-    void DGEMMModelRegion::run(void)
-    {
-        if (m_big_o != 0.0) {
-            if (m_verbosity) {
-                std::cout << "Executing " << m_matrix_size << " x " << m_matrix_size << " DGEMM "
-                          << m_num_progress_updates << " times." << std::endl << std::flush;
-            }
-            ModelRegion::region_enter();
-            for (uint64_t i = 0; i < m_num_progress_updates; ++i) {
-                ModelRegion::loop_enter(i);
-
-                int M = m_matrix_size;
-                int N = m_matrix_size;
-                int K = m_matrix_size;
-                int LDA = m_matrix_size + m_pad_size / sizeof(double);
-                int LDB = m_matrix_size + m_pad_size / sizeof(double);
-                int LDC = m_matrix_size + m_pad_size / sizeof(double);
-                double alpha = 2.0;
-                double beta = 3.0;
-                char transa = 'n';
-                char transb = 'n';
-
-                dgemm(&transa, &transb, &M, &N, &K, &alpha,
-                      m_matrix_a, &LDA, m_matrix_b, &LDB, &beta, m_matrix_c, &LDC);
-
-                ModelRegion::loop_exit();
-            }
-            ModelRegion::region_exit();
-        }
-    }
-
-    StreamModelRegion::StreamModelRegion(double big_o_in, int verbosity, bool do_imbalance, bool do_progress, bool do_unmarked)
-        : ModelRegion(verbosity)
-        , m_array_a(NULL)
-        , m_array_b(NULL)
-        , m_array_c(NULL)
-        , m_array_len(0)
-        , m_align(64)
-    {
-        m_name = "stream";
-        m_do_imbalance = do_imbalance;
-        m_do_progress = do_progress;
-        m_do_unmarked = do_unmarked;
-        big_o(big_o_in);
-        int err = ModelRegion::region(GEOPM_REGION_HINT_MEMORY);
-        if (err) {
-            throw Exception("StreamModelRegion::StreamModelRegion()",
-                            err, __FILE__, __LINE__);
-        }
-    }
-
-    StreamModelRegion::~StreamModelRegion()
-    {
-        big_o(0);
-    }
-
-    void StreamModelRegion::big_o(double big_o_in)
-    {
-        if (m_big_o && m_big_o != big_o_in) {
-            free(m_array_c);
-            free(m_array_b);
-            free(m_array_a);
-        }
-
-        num_progress_updates(big_o_in);
-
-        m_array_len = (size_t)(5e8 * big_o_in);
-        if (big_o_in && m_big_o != big_o_in) {
-            int err = posix_memalign((void **)&m_array_a, m_align, m_array_len * sizeof(double));
-            if (!err) {
-                err = posix_memalign((void **)&m_array_b, m_align, m_array_len * sizeof(double));
-            }
-            if (!err) {
-                err = posix_memalign((void **)&m_array_c, m_align, m_array_len * sizeof(double));
-            }
-            if (err) {
-                throw Exception("StreamModelRegion::big_o(): posix_memalign() failed",
-                                err, __FILE__, __LINE__);
-            }
-#pragma omp parallel for
-            for (size_t i = 0; i < m_array_len; i++) {
-                m_array_a[i] = 0.0;
-                m_array_b[i] = 1.0;
-                m_array_c[i] = 2.0;
-            }
-        }
-        m_big_o = big_o_in;
-    }
-
-    void StreamModelRegion::run(void)
-    {
-        if (m_big_o != 0.0) {
-            if (m_verbosity) {
-                std::cout << "Executing " << m_array_len * m_num_progress_updates << " array length stream triadd."  << std::endl << std::flush;
-            }
-            ModelRegion::region_enter();
-            size_t block_size = m_array_len / m_num_progress_updates;
-            double scalar = 3.0;
-            for (uint64_t i = 0; i < m_num_progress_updates; ++i) {
-                ModelRegion::loop_enter(i);
-#pragma omp parallel for
-                for (size_t j = 0; j < block_size; ++j) {
-                    m_array_a[i * block_size + j] = m_array_b[i * block_size + j] + scalar * m_array_c[i * block_size + j];
-                }
-
-                ModelRegion::loop_exit();
-            }
-            size_t remainder = m_array_len;
-            if (block_size != 0) {
-                remainder = m_array_len % block_size;
-            }
-            for (uint64_t j = 0; j < remainder; ++j) {
-                m_array_a[m_num_progress_updates * block_size + j] = m_array_b[m_num_progress_updates * block_size + j] + scalar * m_array_c[m_num_progress_updates * block_size + j];
-            }
-            ModelRegion::region_exit();
-        }
-    }
-
-    All2allModelRegion::All2allModelRegion(double big_o_in, int verbosity, bool do_imbalance, bool do_progress, bool do_unmarked)
-        : ModelRegion(verbosity)
-        , m_send_buffer(NULL)
-        , m_recv_buffer(NULL)
-        , m_num_send(0)
-        , m_num_rank(0)
-        , m_align(64)
-        , m_rank(-1)
-    {
-        m_name = "all2all";
-        m_do_imbalance = do_imbalance;
-        m_do_progress = do_progress;
-        m_do_unmarked = do_unmarked;
-        big_o(big_o_in);
-        int err = ModelRegion::region(GEOPM_REGION_HINT_UNKNOWN);
-        if (!err) {
-            err = MPI_Comm_rank(MPI_COMM_WORLD, &m_rank);
-        }
-        if (err) {
-            throw Exception("All2allModelRegion::All2allModelRegion()",
-                            err, __FILE__, __LINE__);
-        }
-    }
-
-    All2allModelRegion::~All2allModelRegion()
-    {
-        big_o(0);
-    }
-
-    void All2allModelRegion::big_o(double big_o_in)
-    {
-        if (m_big_o && m_big_o != big_o_in) {
-            free(m_recv_buffer);
-            free(m_send_buffer);
-        }
-
-        num_progress_updates(big_o_in);
-
-        int err = MPI_Comm_size(MPI_COMM_WORLD, &m_num_rank);
-        if (err) {
-            throw Exception("MPI_Comm_size()",
-                            err, __FILE__, __LINE__);
-        }
-
-        if (m_num_progress_updates > 1) {
-            m_num_send = 1048576; //1MB
-        }
-        else {
-            m_num_send = 10485760; //10MB
-        }
-
-        if (big_o_in && m_big_o != big_o_in) {
-            err = posix_memalign((void **)&m_send_buffer, m_align,
-                                 m_num_rank * m_num_send * sizeof(char));
-            if (!err) {
-                err = posix_memalign((void **)&m_recv_buffer, m_align,
-                                     m_num_rank * m_num_send * sizeof(char));
-            }
-            if (err) {
-                throw Exception("All2allModelRegion::big_o(): posix_memalign() failed",
-                                err, __FILE__, __LINE__);
-            }
-#pragma omp parallel for
-            for (size_t i = 0; i < m_num_rank * m_num_send; i++) {
-                m_send_buffer[i] = (char)(i);
-            }
-        }
-        m_big_o = big_o_in;
-    }
-
-    void All2allModelRegion::run(void)
-    {
-        if (m_big_o != 0) {
-            if (m_verbosity) {
-                std::cout << "Executing " << m_num_send << " byte buffer all2all "
-                          << m_num_progress_updates << " times."  << std::endl << std::flush;
-            }
-            MPI_Barrier(MPI_COMM_WORLD);
-            ModelRegion::region_enter();
-            for (uint64_t i = 0; i < m_num_progress_updates; ++i) {
-                ModelRegion::loop_enter(i);
-
-                double timeout = 0.0;
-                struct geopm_time_s start = {{0,0}};
-                struct geopm_time_s curr = {{0,0}};
-                int loop_done = 0;
-                if (!m_rank) {
-                    (void)geopm_time(&start);
-                }
-                while (!loop_done) {
-                    int err = MPI_Alltoall(m_send_buffer, m_num_send, MPI_CHAR, m_recv_buffer,
-                                           m_num_send, MPI_CHAR, MPI_COMM_WORLD);
-                    if (err) {
-                        throw Exception("MPI_Alltoall()", err, __FILE__, __LINE__);
-                    }
-                    if (!m_rank) {
-                        (void)geopm_time(&curr);
-                        timeout = geopm_time_diff(&start, &curr);
-                        if (timeout > (m_big_o / m_num_progress_updates)) {
-                            loop_done = 1;
-                        }
-                    }
-                    err = MPI_Bcast((void*)&loop_done, 1, MPI_INT, 0, MPI_COMM_WORLD);
-                    if (err) {
-                        throw Exception("MPI_Bcast()", err, __FILE__, __LINE__);
-                    }
-                }
-
-                ModelRegion::loop_exit();
-            }
-            ModelRegion::region_exit();
-        }
-    }
-
-    NestedModelRegion::NestedModelRegion(double big_o_in, int verbosity, bool do_imbalance, bool do_progress, bool do_unmarked)
-        : ModelRegion(verbosity)
-        , m_spin_region(big_o_in, verbosity, do_imbalance, do_progress, do_unmarked)
-        , m_all2all_region(big_o_in, verbosity, do_imbalance, do_progress, do_unmarked)
-    {
-    }
-
-    NestedModelRegion::~NestedModelRegion()
-    {
-    }
-
-    void NestedModelRegion::big_o(double big_o_in)
-    {
-        m_spin_region.big_o(big_o_in);
-        m_all2all_region.big_o(big_o_in);
-    }
-
-    void NestedModelRegion::run(void)
-    {
-        if (m_spin_region.m_big_o != 0.0 && m_all2all_region.m_big_o != 0.0) {
-            (void)geopm_prof_epoch();
-        }
-        // Do spin
-        if (m_spin_region.m_big_o != 0.0) {
-            if (m_spin_region.m_verbosity) {
-                std::cout << "Executing " << m_spin_region.m_big_o << " second spin."  << std::endl << std::flush;
-            }
-            (void)geopm_prof_enter(m_spin_region.m_region_id);
-            for (uint64_t i = 0 ; i < m_spin_region.m_num_progress_updates; ++i) {
-                if (m_spin_region.m_do_imbalance) {
-                    (void)geopm_imbalancer_enter();
-                }
-
-                double timeout = 0.0;
-                struct geopm_time_s start = {{0,0}};
-                struct geopm_time_s curr = {{0,0}};
-                (void)geopm_time(&start);
-                while (timeout < m_spin_region.m_delay) {
-                    (void)geopm_time(&curr);
-                    timeout = geopm_time_diff(&start, &curr);
-                }
-
-                if (m_spin_region.m_do_imbalance) {
-                    (void)geopm_imbalancer_exit();
-                }
-            }
-        }
-
-        // Do all2all before spin region exit
-        if (m_all2all_region.m_big_o != 0) {
-            if (m_all2all_region.m_verbosity) {
-                std::cout << "Executing " << m_all2all_region.m_num_send << " byte buffer all2all "
-                          << m_all2all_region.m_num_progress_updates << " times."  << std::endl << std::flush;
-            }
-            for (uint64_t i = 0; i < m_all2all_region.m_num_progress_updates; ++i) {
-                if (m_all2all_region.m_do_imbalance) {
-                    (void)geopm_imbalancer_enter();
-                }
-
-                int err = MPI_Alltoall(m_all2all_region.m_send_buffer, m_all2all_region.m_num_send,
-                                       MPI_CHAR, m_all2all_region.m_recv_buffer,
-                                       m_all2all_region.m_num_send, MPI_CHAR, MPI_COMM_WORLD);
-                if (err) {
-                    throw Exception("MPI_Alltoall()", err, __FILE__, __LINE__);
-                }
-                err = MPI_Barrier(MPI_COMM_WORLD);
-                if (err) {
-                    throw Exception("MPI_Barrier()", err, __FILE__, __LINE__);
-                }
-            }
-
-            if (m_all2all_region.m_do_imbalance) {
-                (void)geopm_imbalancer_exit();
-            }
-        }
-
-        // Do spin part deux.
-        if (m_spin_region.m_big_o != 0.0) {
-            if (m_spin_region.m_verbosity) {
-                std::cout << "Executing " << m_spin_region.m_big_o << " second spin #2."  << std::endl << std::flush;
-            }
-            for (uint64_t i = 0 ; i < m_spin_region.m_num_progress_updates; ++i) {
-                if (m_spin_region.m_do_imbalance) {
-                    (void)geopm_imbalancer_enter();
-                }
-
-                double timeout = 0.0;
-                struct geopm_time_s start = {{0,0}};
-                struct geopm_time_s curr = {{0,0}};
-                (void)geopm_time(&start);
-                while (timeout < m_spin_region.m_delay) {
-                    (void)geopm_time(&curr);
-                    timeout = geopm_time_diff(&start, &curr);
-                }
-
-                if (m_spin_region.m_do_imbalance) {
-                    (void)geopm_imbalancer_exit();
-                }
-            }
-            (void)geopm_prof_exit(m_spin_region.m_region_id);
-        }
-    }
-
-    IgnoreModelRegion::IgnoreModelRegion(double big_o_in, int verbosity, bool do_imbalance, bool do_progress, bool do_unmarked)
-        : ModelRegion(verbosity)
-    {
-        m_name = "ignore";
-        m_do_imbalance = do_imbalance;
-        m_do_progress = do_progress;
-        m_do_unmarked = do_unmarked;
-        big_o(big_o_in);
-        int err = ModelRegion::region(GEOPM_REGION_HINT_IGNORE);
-        if (err) {
-            throw Exception("IgnoreModelRegion::IgnoreModelRegion()",
-                            err, __FILE__, __LINE__);
-        }
-    }
-
-    IgnoreModelRegion::~IgnoreModelRegion()
-    {
-
-    }
-
-    void IgnoreModelRegion::big_o(double big_o_in)
-    {
-        num_progress_updates(big_o_in);
-        double seconds = big_o_in / m_num_progress_updates;
-        m_delay = {(time_t)(seconds),
-                   (long)((seconds - (time_t)(seconds)) * 1E9)};
-
-        m_big_o = big_o_in;
-    }
-
-    void IgnoreModelRegion::run(void)
-    {
-        if (m_big_o != 0.0) {
-            if (m_verbosity) {
-                std::cout << "Executing ignored " << m_big_o << " second sleep."  << std::endl << std::flush;
-            }
-            ModelRegion::region_enter();
-            for (uint64_t i = 0 ; i < m_num_progress_updates; ++i) {
-                ModelRegion::loop_enter(i);
-
-                int err;
-                err = clock_nanosleep(CLOCK_REALTIME, 0, &m_delay, NULL);
-                if (err) {
-                    throw Exception("IgnoreModelRegion::run()",
-                                    GEOPM_ERROR_RUNTIME, __FILE__, __LINE__);
-                }
-
-                ModelRegion::loop_exit();
-            }
-            ModelRegion::region_exit();
         }
     }
 }
