+++ test_integration/Makefile.mk
@@ -49,20 +49,17 @@ noinst_PROGRAMS += test_integration/test_ee_stream_dgemm_mix \
                    # end
 
 test_integration_test_ee_stream_dgemm_mix_SOURCES = test_integration/test_ee_stream_dgemm_mix.cpp \
-                                                    src/ModelRegion.cpp \
-                                                    src/ModelRegion.hpp \
+                                                    $(model_source_files) \
                                                     # end
 test_integration_test_ee_stream_dgemm_mix_LDADD = libgeopm.la $(MATH_LIB) $(MPI_CLIBS)
 test_integration_test_ee_stream_dgemm_mix_LDFLAGS = $(AM_LDFLAGS) $(MPI_CLDFLAGS) $(MATH_CLDFLAGS)
 test_integration_test_ee_stream_dgemm_mix_CXXFLAGS = $(AM_CXXFLAGS) $(MPI_CFLAGS) -D_GNU_SOURCE -std=c++11 $(MATH_CFLAGS)
 
-if ENABLE_OPENMP
 noinst_PROGRAMS += test_integration/test_omp_outer_loop \
                    # end
 
 test_integration_test_omp_outer_loop_SOURCES = test_integration/test_omp_outer_loop.cpp \
-                                               src/ModelRegion.cpp \
-                                               src/ModelRegion.hpp \
+                                               $(model_source_files) \
                                                # end
 test_integration_test_omp_outer_loop_LDADD = libgeopm.la $(MATH_LIB) $(MPI_CLIBS)
 test_integration_test_omp_outer_loop_LDFLAGS = $(AM_LDFLAGS) $(MPI_CLDFLAGS) $(MATH_CLDFLAGS)
